requirements.txt
melotts==0.1.1
openai-whisper